Move cli stuff to __main__, improve argparser, add docs
