+++ ldap_sync/__init__.py
@@ -1,0 +1,6 @@
+"""
+ldap_sync
+
+This package provides a standalone LDAP syncer.  For more information
+on how to execute it, run ``python -m ldap_sync --help``.
+"""
